fix: botocore client error
An error occurred (MalformedXML) when calling the DeleteObjects
operation: The XML you provided was not well-formed or did not validate
against our published schema.

+++ back/session/storage/async_s3.py
@@ -268,6 +268,9 @@ async def delete(client, bucket_name: str, prefix: str) -> bool:
         for p in _page.CommonPrefixes:
             await delete(client, bucket_name, p.Prefix)
 
+        if len(objects) == 0:
+            continue
+
         resp = await client.delete_objects(
             Bucket=bucket_name, Delete={"Objects": objects}
         )
